fix admin code mixup bug
